HADOOP-1498.  Replace boxed types with primitives in many places.  Contributed by stack.
git-svn-id: https://svn.apache.org/repos/asf/lucene/hadoop/trunk@548523 13f79535-47bb-0310-9956-ffa450edef68

+++ src/contrib/hbase/src/java/org/apache/hadoop/hbase/HServerAddress.java
@@ -45,7 +45,8 @@ public class HServerAddress implements Writable {
       throw new IllegalArgumentException("Not a host:port pair: " + hostAndPort);
     }
     String host = hostAndPort.substring(0, colonIndex);
-    int port = Integer.valueOf(hostAndPort.substring(colonIndex + 1));
+    int port =
+      Integer.valueOf(hostAndPort.substring(colonIndex + 1)).intValue();
     this.address = new InetSocketAddress(host, port);
     this.stringValue = hostAndPort;
   }
